fix: console error on user profile page
