--- js/src/forum/extenders/index.ts
@@ -1,9 +1,0 @@
-import SettingsPage from './SettingsPage';
-import Discussion from './Discussion';
-import User from './User';
-
-export default () => {
-  Discussion();
-  SettingsPage();
-  User();
-};
