+++ js/src/forum/index.ts
@@ -1,13 +1,18 @@
 import app from 'flarum/forum/app';
-import extenders from './extenders';
 import pages from './pages';
-import extendNotificationGrid from './notifications/extendNotificationGrid';
+import extendNotificationGrid from './extenders/extendNotificationGrid';
+import extendUserComponents from './extenders/extendUserComponents';
+import extendSettingsPage from './extenders/extendSettingsPage';
+import Discussion from './extenders/Discussion';
 
 export { default as extend } from './extend';
 
 app.initializers.add('fof-byobu', function () {
-  extenders();
+  Discussion();
 
   pages();
   extendNotificationGrid();
+
+  extendUserComponents();
+  extendSettingsPage();
 });
